SumEvenFibonacci
This program evaluates the sum of the first four million even numbers of the Fibonacci sequence. This is not created by using a recurrence function, so I think it is quite inefficient. I may be looking into a better way to compute this sum using a recurrence or some other function.